data_sync/requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
nba_api>=1.4.0
